requirements.txt
# Core Data Processing
pyspark==3.5.0
kafka-python==2.0.2
pandas==2.1.3
numpy==1.26.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4

# Visualization
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0

# Data Generation & Utilities
faker==20.1.0
python-dateutil==2.8.2

# Configuration & Logging
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# API (for future extensions)
fastapi==0.104.1
uvicorn==0.24.0


